requirements.txt
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.30.0
Office365-REST-Python-Client==2.5.11
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
